Extract shared utilities from read_jleague_matches.py into match_utils.py
Separate competition-agnostic functions (CSV I/O, timestamp management,
DataFrame comparison, date parsing, CLI arg parsing) from J-League-specific
scraping logic. Update all importers to use match_utils directly.

Fixes #87
